+++ test/unit/test_helper_test.rb
@@ -1,4 +1,4 @@
-require 'helper'
+require_relative '../helper'
 require 'focused_controller/test_helper'
 require 'action_controller'
 
@@ -23,13 +23,11 @@ module FocusedController
 
         def self._routes
           OpenStruct.new(
-            :named_routes => OpenStruct.new(
-              :module => Module.new do
-                def foo_path
-                  '/foo'
-                end
+            :url_helpers => Module.new do
+              def foo_path
+                '/foo'
               end
-            )
+            end
           )
         end
       end
@@ -37,13 +35,11 @@ module FocusedController
       class Show < Action
         def self._routes
           OpenStruct.new(
-            :named_routes => OpenStruct.new(
-              :module => Module.new do
-                def bar_path
-                  '/bar'
-                end
+            :url_helpers => Module.new do
+              def bar_path
+                '/bar'
               end
-            )
+            end
           )
         end
       end
@@ -177,38 +173,6 @@ module FocusedController
         controller.call
 
         subject.cookies[:foo].must_equal 'omg'
-      end
-
-      describe 'url stubbing' do
-        before do
-          subject.stub_url :foo, :bar
-        end
-
-        it 'handles equality' do
-          foo, bar = Object.new, Object.new
-          StubbedURL.new(:foo_path, [foo]).must_equal StubbedURL.new(:foo_path, [foo])
-          StubbedURL.new(:foo_path, [foo, bar]).wont_equal StubbedURL.new(:foo_path, [foo])
-          StubbedURL.new(:foo_path, [bar]).wont_equal StubbedURL.new(:foo_path, [foo])
-          StubbedURL.new(:bar_path, [foo]).wont_equal StubbedURL.new(:foo_path, [foo])
-        end
-
-        it 'has a to_s' do
-          StubbedURL.new(:foo_path, ['omg', 'lol']).to_s.must_equal "foo_path(omg, lol)"
-        end
-
-        it 'creates a stub method on the test and controller instances' do
-          foo = Object.new
-          subject.foo_path(foo).must_equal StubbedURL.new(:foo_path, [foo])
-          subject.controller.foo_path(foo).must_equal StubbedURL.new(:foo_path, [foo])
-          subject.foo_url(foo).must_equal StubbedURL.new(:foo_url, [foo])
-          subject.controller.foo_url(foo).must_equal StubbedURL.new(:foo_url, [foo])
-        end
-
-        it 'works with a controller' do
-          foo = Object.new
-          subject.controller.redirect_to StubbedURL.new(:foo_path, [foo])
-          must_succeed { subject.assert_redirected_to StubbedURL.new(:foo_path, [foo]) }
-        end
       end
     end
   end