+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+## 2.0.3 ##
+
+* Fix RSpec 3 deprecation warning
+
+## 2.0.2 ##
+
+* Rails 4.2.5.1 compatibility fix
+
+## 2.0.1 ##
+
+* Routing bug fix
+
+## 2.0.0 ##
+
+* Rails 3.X and 4.0 no longer supported
+* Ruby 1.9 no longer officially supported (it may still work)
+* Isolated unit tests (via `stub_url`) no longer supported. Use the
+  Spring preloader to keep your app warm instead.
+* Rails 4.2 support added
+
 ## Version 1.2.1 ##
 
 * Fix gemspec